requirements.txt
flask
gunicorn
psycopg2-binary
requests
Flask-gzip